+++ packages/azure/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @ai-sdk/azure
 
+## 1.0.15
+
+### Patch Changes
+
+- Updated dependencies [b19aa82]
+  - @ai-sdk/openai@1.0.13
+
+## 1.0.14
+
+### Patch Changes
+
+- Updated dependencies [a4241ff]
+  - @ai-sdk/openai@1.0.12
+
 ## 1.0.13
 
 ### Patch Changes